creating a tavi high-level interface
