+++ src/tavi/tavi_model/tavi_project.py
@@ -1,86 +1,46 @@
 import os
-from dataclasses import field
-from typing import Any, Iterable, Optional
-
-import numpy as np
+from typing import Iterable, Optional
 
 from tavi.tavi_model.FileSystem.load_manager import LoadManager
-from tavi.tavi_model.FileSystem.tavi_class_factory import Scan
 from tavi.tavi_model.filter import Filter, Logic, Operations
+from tavi.tavi_model.tavi_data import TaviData
 
 
 class TaviProject:
     """
-    A container class for managing a complete TAVI project, including
-    raw TAS (triple-axis spectrometer) experimental data, processed data,
-    filtering operations, fitting, and plotting management.
+    Central container for managing a full TAVI project workflow.
 
-    The `TaviProject` serves as the central data structure for organizing
-    experimental scans and coordinating workflows such as loading, filtering,
-    combining, fitting, and visualizing data. It acts as a high-level interface
-    to link raw data, metadata, and analysis pipelines in TAVI.
+    The `TaviProject` class provides a high-level interface for loading,
+    filtering, processing, and analyzing triple-axis spectrometer (TAS) data.
+    It acts as the core object for managing raw scans, metadata, selected data,
+    and later processing steps such as combining, fitting, and plotting.
+
+    Internally, `TaviProject` owns a `TaviData` object, which stores and
+    organizes all scan-related information (raw and processed).
+
+    Typical usage involves:
+        1. Loading raw scan files into the project with `load_scans`.
+        2. Selecting subsets of scans or points with `select_scans`.
+        3. Performing transformations such as combining or fitting.
+        4. Visualizing or exporting results.
 
     Attributes:
-        scans (dict[str, Scan]):
-            A mapping from scan identifiers (e.g., filenames or scan numbers)
-            to their corresponding `Scan` objects, which hold raw data and metadata.
+        tavi_data (TaviData):
+            The main data container that stores scan lists, processed data,
+            and user selections (for viewing or modeling).
 
-        combined_data (dict[str, np.ndarray]):
-            Stores merged or aggregated data from multiple scans for analysis.
-
-        filtered_data (dict[str, np.ndarray]):
-            Stores scan data after user-applied filters (e.g., Q/E cuts, masks).
-
-        view_selected_data (dict[str, Any]):
-            Holds user-selected data for visualization purposes.
-
-        process_selected_data (dict[str, Any]):
-            Holds user-selected data for processing workflows.
-
-        fit_manager (dict[str, Any]):
-            A container for fitting models, results, and related configurations.
-
-        plot_manager (dict[str, Any]):
-            A container for managing plotting sessions, styles, and figures.
-
-    Methods:
-        load_scans(data_folder, data_files, ub_dir=None, facility=None):
-            Load scans from disk using the `LoadManager`. Populates `self.scans`.
-
-        load_tavi():
-            (Placeholder) Load a serialized TAVI project file.
-
-        save_tavi():
-            (Placeholder) Save the current TAVI project to disk.
-
-        filter_scans():
-            (Placeholder) Apply filters to raw scan data and store results.
-
-        combine_data():
-            (Placeholder) Combine multiple scans into `combined_data`.
-
-        fit_data():
-            (Placeholder) Perform fitting routines on selected datasets.
-
-        plot_data():
-            (Placeholder) Generate plots of scan or processed data.
+    Notes:
+        - The `load_scans` method uses a `LoadManager` backend for parsing
+          facility-specific file formats (e.g., ORNL, ILL, ISIS).
+        - Scan filtering is handled by the `Filter` class with logical
+          conditions and operations defined in `tavi_model.filter`.
+        - Future methods (`load_tavi`, `save_tavi`, `combine_data`,
+          `fit_data`, `plot_data`) will extend project persistence,
+          advanced analysis, and visualization.
     """
 
-    def __init__(
-        self,
-        scans: dict[str, Scan] = field(default_factory=dict),
-        combined_data: dict[str, np.array] = field(default_factory=dict),
-        process_selected_data: list[str] = [],  # mouse selection
-        show_selected_data: list[str] = [],  # display
-        fit_manager: dict[str, Any] = field(default_factory=dict),
-        plot_manager: dict[str, Any] = field(default_factory=dict),
-    ):
-        self.scans = scans
-        self.combined_data = combined_data
-        self.process_selected_data = process_selected_data
-        self.show_selected_data = show_selected_data
-        self.fit_manager = fit_manager
-        self.plot_manager = plot_manager
+    def __init__(self):
+        self.tavi_data = TaviData()
 
     # --------------------Load Manager-------------------------------------
     def load_scans(
@@ -97,7 +57,7 @@ class TaviProject:
         from the specified directory and populate the `scans` attribute with
         `Scan` objects containing raw data, metadata, and UB matrix information.
 
-        Args:
+        Attributes:
             data_folder (os.PathLike | str, optional):
                 Path to the directory containing the scan files.
                 If not provided, defaults to the current working directory.
@@ -118,7 +78,7 @@ class TaviProject:
             None
                 The method updates the `scans` attribute in place.
         """
-        self.scans = LoadManager(
+        self.tavi_data.scan_list = LoadManager(
             data_folder=data_folder, data_files=data_files, ub_dir=ub_dir, facility=facility
         ).load()
 
@@ -138,14 +98,14 @@ class TaviProject:
         category: Optional[str] = None,
         tol=0.01,
     ) -> None:
-        filtered_data = Filter(self.scans, conditions=conditions, and_or=and_or).filter_data()
+        filtered_data = Filter(self.tavi_data.scan_list, conditions=conditions, and_or=and_or).filter_data()
         match category:
             case "view":
-                self.view_selected_data.append(filtered_data)
+                self.tavi_data.show_selected_data.append(filtered_data)
             case "model":
-                self.process_selected_data.append(filtered_data)
+                self.tavi_data.process_selected_data = filtered_data
             case _:
-                self.process_selected_data.append(filtered_data)
+                self.tavi_data.show_selected_data.append(filtered_data)
 
     # TO DO
     def combine_data():
@@ -168,11 +128,11 @@ if __name__ == "__main__":
 
     TaviProj.load_scans(filepath)
 
-    # filename = "CG4C_exp0424_scan0042.dat"
-    TaviProj.select_scans(
-        conditions=([["scan", Operations.CONTAINS, "42"], ["scan", Operations.CONTAINS, "4"]]), and_or=Logic.OR
-    )
-    print(TaviProj.process_selected_data)
+    filename = "CG4C_exp0424_scan0042.dat"
+    TaviProj.select_scans(conditions=([["scan", Operations.CONTAINS, "42"]]), and_or=Logic.OR)
+
+    TaviProj.select_scans(conditions=([["scan", Operations.CONTAINS, "4"]]), and_or=Logic.OR)
+    print(TaviProj.tavi_data.show_selected_data)
     # print(type(TaviProj.scans[filename].metadata.scan))
 #     print(TaviProj.scans[filename].ubconf)
 # print(TaviProj.scans[filename].data.Pt)
